:construction: Use avatar and avatar_provider in various places
This patch starts preparing the user interface for supporting multiple
avatar providers by adding the ability to edit these values in the edit
form and updating the user overview page.

It also introduces an avatar provider enum currently only containing
gravatar.

As the current API version implement still expects an email field
we return email instead of avatar and avatar_provider and only have
it be non null when avatar_provider is set to gravatar. Likewise
passing an email in update or create call sets avatar to the email
and avatar_provider to gravatar.