:sparkles: Add WebFinger / ActivityPub avatar provider
This patch introduces a new avatar provider based on the changes from
the previous commits.

The WebFinger / ActivityPub provider ("webfinger" internally)
loosely interprets the avatar field in a user as an RFC 7565 acct URI,
the library used also allows only specifiying a domain name though.

Based on this it will contact the host and query the WebFinger endpoint.

If the WebFinger response contains an avatar relation, the link contained
in that relation will used.

Only some fediverse software supports this however, as such we also
implement a tiny bit of ActivityPub and directly query the user using
its "application/activity+json" relation.

Given a Mastodon url (and optionally an access token), mete can also
use the Mastodon client api to resolve the actor returned by WebFinger,
which is generally more reliable than speaking ActivityPub directly, as
requests can be signed by the Mastodon server.

The results of these lookups will be cached for one day and can be
invalidated by updating the user from the UI.